echome/requirements.txt
bcrypt==3.1.7
cffi==1.14.0
click==7.1.2
cryptography==2.9.2
ecdsa==0.15
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
libvirt-python==6.3.0
MarkupSafe==1.1.1
pkg-resources==0.0.0
psycopg2==2.8.5
pycparser==2.20
PyJWT==1.7.1
PyYAML==5.3.1
six==1.15.0
SQLAlchemy==1.3.17
sshpubkeys==3.1.0
uWSGI==2.0.18
uwsgidecorators==1.1.0
Werkzeug==1.0.1
